Move processing into cache
